Add asserts for incorrect ICF reg deps to the Power codegen
Previously, we encountered a number of difficult-to-diagnose issues
relating to missing or incorrect register dependencies on end of
internal control flow labels in the Power codegen. The most prevalent
cause of such issues is a missing r0 exclude on a register dependency.
If the register for a dependency in ICF happened to be allocated to r0
and was used where r0 has special meaning (i.e. as a base register or in
an addi instruction), the local register allocator would shuffle
registers in the middle of ICF, causing numerous issues.

To prevent such issues from being introduced accidentally in the future,
a couple of new asserts have been added that ensure that all necessary
r0 excludes are present inside ICF. These asserts are designed to fire
if the register in question *could* have been assigned to r0 (even if it
actually wasn't) to allow such issues to be caught as early as possible.

Signed-off-by: Ben Thomas <ben@benthomas.ca>