Remove ICF labels on ZEROCHK code for Power
Previously, mainline code for ZEROCHK nodes in the Power codegen was
being marked as internal control flow. However, this was triggering the
new ICF-related asserts in the Power codegen since register dependencies
were not set up correctly. Since all currently known users of ZEROCHK
use helper calls that don't return, this isn't necessarily an issue in
practice. For this reason, the code is being temporarily unmarked as ICF
to silence the assert.

Signed-off-by: Ben Thomas <ben@benthomas.ca>