Refactor the Power compare evaluators
The Power codegen's evaluators for the integer and float compare opcodes
have been significantly refactored. This refactoring is in preparation
for the addition of a fast path for evaluating compares directly to
condition registers in circumstances where the result of a comparison
node would need to be compared against 0, e.g. when evaluating select
nodes.

Additionally, the new versions have a new optimization when comparing
values of sub-register types: extend the values, subtract them, then
check the sign bit of the result. This avoids having to perform the
usual branch sequence with only 3-5 instructions.

Signed-off-by: Ben Thomas <ben@benthomas.ca>