Disable the pTOC under POWER10
Since POWER10 processors support PC-relative addressing for prefixed
loads, the pTOC is now redundant and all pTOC accesses can be replaced
with PC-relative loads. In line with this, the pTOC will be disabled
when POWER10 support is enabled and will always appear to be full, just
in case any uses of it remain.

Co-authored-by: Julian Wang <zlwang@ca.ibm.com>
Signed-off-by: Ben Thomas <ben@benthomas.ca>