Fix addToOffset expansion for POWER10
Previously, the MemoryReference::addToOffset method would automatically
expand out-of-range displacements using the addis instruction. However,
when POWER10 support is enabled, it is generally preferable to use the
prefixed version of the load or store instead. Since the transformation
that causes this exploitation to occur happens in instruction expansion,
we must leave displacements that are out of the normal 16-bit range but
within the 34-bit range alone at this stage for exploitation of the
prefixed loads and stores to occur.

Signed-off-by: Ben Thomas <ben@benthomas.ca>