Add 34-bit displacement exploitation in accessStaticItem
Previously, the MemoryReference::accessStaticItem method would always
use ConstantDataSnippet when POWER10 support was enabled. While this is
better than trying to use the pTOC, the prefixed load and store
instructions also provide the ability to to use 34-bit displacements,
which can be used to load from absolute addresses if they can be
represented fully in this displacement. Exploitation of this has now
been added to avoid the extra load.

Signed-off-by: Ben Thomas <ben@benthomas.ca>