Fix loadaddrEvaluator support for null base register
The previous implementation of loadaddrEvaluator in the Power codegen
assumed that the constructed MemoryReference would never have a base
register of NULL. However, exploitation of POWER10 prefixed loads and
stores will make use of a NULL base register to perform loads to
addresses that can be represented as a 34-bit signed integer. To support
this, loadaddrEvaluator has been updated to handle this case.

Signed-off-by: Ben Thomas <ben@benthomas.ca>