Add POWER9 setb exploitation to compare evaluators
Prior to the original refactor of the Power codegen's compare
evaluators, [fd]cmplt and [fd]cmpgt would exploit the setb instruction
available on POWER9 to compute their values without branching. Although
this does require some extra instructions to get the correct result from
the 3-way setb, it avoids the branching sequence and so should eliminate
the cost of branch misprediction there.

As it turns out, a similar approach can be applied to all compares. This
allows the result of any compare to be placed in a register in 4 or
fewer instructions (plus the compare itself). This should provide a
performance boost when the compare opcodes are used on POWER9 machines.

Signed-off-by: Ben Thomas <ben@benthomas.ca>